[kyne.k] 방탈출 인증 미션 Step 1,2 리뷰 요청합니다! (#42)
* docs: 프로그램 요구사항 정리

* feat: 토큰 발급 API 생성

* docs: 구현된 기능 문서 반영

* feat: 내 정보 조회하기 구현

* feat: 예약시 로그인 여부 확인 기능 추가

* docs: 예약시 로그인 여부 확인기능 추가

* feat: 비로그인된 사용자는 예약 불가하도록 변경

* feat: 예약 취소시 회원 로그인 상태 및 권한 확인 기능 추가

* refactor : 불필요한 테스트 코드, 로깅 제거 / 권한 체크 로직 수정 / 예외 처리 방식 변경

* feat : 로그인 검증 기능 생성

* refactor : 롬복 적용 리펙터링

* feat : 로그인 인터셉터 생성

* docs : README.md 구현 목록 작성

+++ src/main/java/nextstep/reservation/ReservationController.java
@@ -1,5 +1,10 @@
 package nextstep.reservation;
 
+import nextstep.member.Member;
+import nextstep.member.MemberService;
+import nextstep.exception.AuthorizationException;
+import nextstep.exception.ForbiddenAccessException;
+import nextstep.ui.AuthenticationPrincipal;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
@@ -9,34 +14,50 @@ import java.util.List;
 @RestController
 @RequestMapping("/reservations")
 public class ReservationController {
-
     public final ReservationService reservationService;
+    public final MemberService memberService;
 
-    public ReservationController(ReservationService reservationService) {
+    public ReservationController(ReservationService reservationService, MemberService memberService) {
         this.reservationService = reservationService;
+        this.memberService = memberService;
     }
 
     @PostMapping
-    public ResponseEntity createReservation(@RequestBody ReservationRequest reservationRequest) {
+    public ResponseEntity<URI> createReservation(@AuthenticationPrincipal String token,
+                                                 @RequestBody ReservationRequest reservationRequest) {
+        if (token == null) {
+            throw new AuthorizationException();
+        }
+        memberService.findByToken(token);
         Long id = reservationService.create(reservationRequest);
         return ResponseEntity.created(URI.create("/reservations/" + id)).build();
     }
 
     @GetMapping
-    public ResponseEntity readReservations(@RequestParam Long themeId, @RequestParam String date) {
+    public ResponseEntity<List<Reservation>> readReservations(@RequestParam Long themeId, @RequestParam String date) {
         List<Reservation> results = reservationService.findAllByThemeIdAndDate(themeId, date);
         return ResponseEntity.ok().body(results);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity deleteReservation(@PathVariable Long id) {
+    public ResponseEntity<Void> deleteReservation(@AuthenticationPrincipal String token, @PathVariable Long id) {
+        if (token == null) {
+            throw new AuthorizationException();
+        }
+
+        Member member = memberService.findByToken(token);
+        Reservation reservation = reservationService.findById(id);
+        if (!reservation.getName().equals(member.getUsername())) {
+            throw new ForbiddenAccessException();
+        }
+
         reservationService.deleteById(id);
 
         return ResponseEntity.noContent().build();
     }
 
-    @ExceptionHandler(Exception.class)
-    public ResponseEntity onException(Exception e) {
+    @ExceptionHandler(NullPointerException.class)
+    public ResponseEntity<Void> onException(Exception e) {
         return ResponseEntity.badRequest().build();
     }
 }
